+++ lib/src/samplers/string.dart
@@ -1,8 +1,10 @@
 import '../sample.dart';
 
+/// Sampler for [String], returning a single-character [String].
 class StringSampler implements Sampler<String, String> {
   @override
   int size(String s) => s.length;
+
   @override
-  String getAt(String s, int i) => s[i];
+  String getAt(String s, int index) => s[index];
 }