+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.0.1
+- Added documentation
+- Added sampleMany extension
+- Refactored codebase
+
 ## 1.0.0
 
 - Initial version.
+- Supports sampling from string, list, map, and iterator types.
+- Supports uniform distribution & weighted distribution